Resolve #393, upgrade Go module to v2
