+++ go.mod
@@ -1,7 +1,8 @@
-module github.com/360EntSecGroup-Skylar/excelize
+module github.com/360EntSecGroup-Skylar/excelize/v2
+
+go 1.12
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826
 	github.com/stretchr/testify v1.3.0
 )
